TruncatedStreams

Documentation for TruncatedStreams.

TruncatedStreams.FixedLengthIO

TruncatedStreams.SentinelIO

TruncatedStreams.TruncatedIO

TruncatedStreams.unwrap

TruncatedStreams.FixedLengthIO — Type

FixedLengthIO(io, length) <: TruncatedIO

A truncated source that reads io up to length bytes.

TruncatedStreams.SentinelIO — Type

SentinelIO(io, sentinel) <: TruncatedIO

A truncated source that reads io until sentinel is found.

Can be reset with the reseteof() method if the sentinel read is discovered upon further inspection to be bogus.

The wrapped stream io must implement mark and reset.

TruncatedStreams.TruncatedIO — Type

TruncatedIO <: IO

Wraps an IO stream object and lies about how much is left to read from the stream.

Objects inheriting from this abstract type pass along all IO methods to the wrapped stream except for bytesavailable(io) and eof(io). Inherited types need to implement:

unwrap(::AbstractTruncatedSource)::IO: return the wrapped IO stream.

Base.bytesavailable(::AbstractTruncatedSource)::Int: report the number of bytes available to

read from the stream until EOF or a buffer refill.

Base.eof(::AbstractTruncatedSource)::Bool: report whether the stream cannot produce any more

bytes.

Optional methods to implement that might be useful are:

Base.reseteof(::AbstractTruncatedSource)::Nothing: reset EOF status.

Base.unsafe_read(::AbstractTruncatedSource, p::Ptr{UInt8}, n::UInt)::Nothing: copy n bytes from the stream into memory pointed to by p.

Base.seek(::TruncatedIO, n::Integer) and Base.seekend(::TruncatedIO): seek stream to position n or end of stream.

Base.reset(::TruncatedIO): reset a marked stream to the saved position.

TruncatedStreams.unwrap — Function

unwrap(s<:TruncatedIO) -> IO

Return the wrapped source.