TruncatedStreams

Documentation for TruncatedStreams.

TruncatedStreams.FixedLengthIO

TruncatedStreams.SentinelIO

TruncatedStreams.TruncatedIO

TruncatedStreams.unwrap

TruncatedStreams.FixedLengthIO — Type

FixedLengthIO(io, length) <: TruncatedIO

A truncated source that reads io up to length bytes.

julia> io = IOBuffer(collect(0x00:0xff));

julia> fio = FixedLengthIO(io, 10);

julia> read(fio)
10-element Vector{UInt8}:
 0x00
 0x01
 0x02
 0x03
 0x04
 0x05
 0x06
 0x07
 0x08
 0x09

julia> eof(fio)
true

As soon as a read from a FixedLengthIO object would read past length bytes of the underlying IO stream, EOF is signalled, potentially leading to an EOFError being thrown.

julia> read(fio, Int)
ERROR: EOFError: read end of file
[...]

Seeking does not affect the length at which the stream is truncated, but may affect how many bytes are available to read.

julia> seek(fio, 8); read(fio)
2-element Vector{UInt8}:
 0x08
 0x09

Writing to a FixedLengthIO object does not affect the length at which the stream is truncated, but may affect how many bytes are available to read.

julia> io = IOBuffer(collect(0x00:0x05); read=true, write=true); fio = FixedLengthIO(io, 10);

julia> read(fio)
6-element Vector{UInt8}:
 0x00
 0x01
 0x02
 0x03
 0x04
 0x05

julia> write(fio, collect(0x06:0xff));

julia> seekstart(fio);  # writing advances the IOBuffer's read pointer

julia> read(fio)
10-element Vector{UInt8}:
 0x00
 0x01
 0x02
 0x03
 0x04
 0x05
 0x06
 0x07
 0x08
 0x09

TruncatedStreams.SentinelIO — Type

SentinelIO(io, sentinel) <: TruncatedIO

A truncated source that reads io until sentinel is found.

julia> io = IOBuffer(collect(0x00:0xff));

julia> sio = SentinelIO(io, [0x0a, 0x0b]);

julia> read(sio)
10-element Vector{UInt8}:
 0x00
 0x01
 0x02
 0x03
 0x04
 0x05
 0x06
 0x07
 0x08
 0x09

julia> eof(sio)
true

As soon as a read from a SentinelIO object would read the start of a byte sequence matching sentinel from the underlying IO stream, EOF is signalled, potentially leading to an EOFError being thrown.

julia> read(sio, Int)
ERROR: EOFError: read end of file
[...]

Seeking does not affect reading of the sentinel, but may affect how many bytes are available to read.

julia> seek(sio, 8); read(sio)
2-element Vector{UInt8}:
 0x08
 0x09

Writing to a SentinelIO object does not affect the length at which the stream is truncated, but may affect how many bytes are available to read.

julia> io = IOBuffer(collect(0x00:0x07); read=true, write=true); sio = SentinelIO(io, [0x06, 0x07]);

julia> read(sio)
6-element Vector{UInt8}:
 0x00
 0x01
 0x02
 0x03
 0x04
 0x05

julia> write(sio, collect(0x01:0xff));

julia> seekstart(sio);  # writing advances the IOBuffer's read pointer

julia> read(sio)  # still the same output because the sentinel is still there
6-element Vector{UInt8}:
 0x00
 0x01
 0x02
 0x03
 0x04
 0x05

Detection of eof can be reset with the Base.reseteof() method. Use this if the sentinel that was read is determined upon further inspection to be bogus.

julia> Base.reseteof(sio)  # that last sentinel was fake, so reset EOF and read again

julia> read(sio)  # returns the first sentinel found and continues to read until the next one is found
7-element Vector{UInt8}:
 0x06
 0x07
 0x01
 0x02
 0x03
 0x04
 0x05

Note

If the wrapped stream does not contain a sentinel, reading to the end of the stream will throw EOFError.

julia> io = IOBuffer(collect(0x00:0x07)); sio = SentinelIO(io, [0xff, 0xfe]);

julia> read(sio)
ERROR: EOFError: read end of file
[...]

TruncatedStreams.TruncatedIO — Type

TruncatedIO <: IO

Wraps a streaming IO object that reads only as much as should be read and not a byte more.

Objects inheriting from this abstract type pass along all IO methods to the wrapped stream except for bytesavailable(io) and eof(io). Inherited types must implement:

TruncatedStreams.unwrap(::TruncatedIO)::IO: return the wrapped IO stream.

Base.eof(::TruncatedIO)::Bool: report whether the stream cannot produce any more bytes.

In order to implement truncation, some number of these methods will likely need to be implemented:

Base.unsafe_read(::TruncatedIO, p::Ptr{UInt8}, n::UInt)::Nothing: copy n bytes from the stream into memory pointed to by p.

Base.read(::TruncatedIO, T::Type)::T: read and return an object of type T from the stream.

Base.bytesavailable(::TruncatedIO)::Int: report the number of bytes available to read from the stream until EOF or a buffer refill.

Base.seek(::TruncatedIO, p::Integer) and Base.seekend(::TruncatedIO): seek stream to position p or end of stream.

Base.reset(::TruncatedIO): reset a marked stream to the saved position.

Base.reseteof(::TruncatedIO)::Nothing: reset EOF status.

Note that writing to the stream does not affect truncation.

The following methods must be implemented by the wrapped IO type for all the functionality of the truncated streams to work at all:

Base.eof(::IO)::Bool

Base.read(::IO, ::Type{UInt8})::UInt8

The wrapped stream also must implement Base.seek and Base.skip for seeking and skipping of the truncated stream to work properly. Additionally, Base.position needs to be implemented for some instances of Base.seek to work properly.

TruncatedStreams.unwrap — Function

unwrap(s<:TruncatedIO) -> IO

Return the wrapped source.